Rename file solver_options.py to solver_option.py
