+++ torchqdynamics/solver.py
@@ -1,5 +1,5 @@
 """Gather all solver options under the same namespace `solver`."""
 
-from .mesolve.solver_options import Rouchon, Rouchon1, Rouchon1_5, Rouchon2
-from .sesolve.solver_options import Propagator
-from .solver_options import Dopri45, Euler
+from .mesolve.solver_option import Rouchon, Rouchon1, Rouchon1_5, Rouchon2
+from .sesolve.solver_option import Propagator
+from .solver_option import Dopri45, Euler
